--- dev-requirements.txt
@@ -1,1 +1,0 @@
-pudb==2019.2
