+++ release_notes/2.1.1.md
@@ -1,0 +1,10 @@
+**EclecticIQ app Release Notes - Published by EclecticIQ October 20, 2023**
+
+
+**Version 2.1.1 - Released October 20, 2023**
+
+* new actions added: search entity by id, make POST, make GET, make DELETE request to EclecticIQ Intelligence Center
+* added support for EclecticIQ Public API v1 and v2
+* default tags for Sightings and Indicators fixed to Splunk SOAR from Phantom
+* default timeout changed to 300s for API calls
+* massive bug fix
