--- requirements.txt
@@ -1,2 +1,0 @@
-beautifulsoup4==4.9.1
-requests==2.25.0
